fix: add 5-second delay after IP Set creation to avoid race condition
AWS WAF was failing with WAFUnavailableEntityException when trying to
create the Web ACL immediately after creating the IP Set. This is because
AWS WAF needs a few seconds for the IP Set to propagate before it can be
referenced in a Web ACL rule.

Added 5-second sleep between IP Set and Web ACL creation to avoid this
race condition.